refactor backend to render index.html from build directory
