--- backend/Pipfile
@@ -1,13 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-djangorestframework = "*"
-django-cors-headers = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.8"
