--- src/components/MentorForm/MentorForm.css
@@ -1,3 +1,0 @@
-.required {
-  content: '*';
-}
